% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function plots the locations of accidents for a given state in a given year on that state's map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number to plot data for}

\item{year}{The year to plot data for}
}
\value{
Nothing. Outputs a plot of the locations of accidents on a state map for a given year
}
\description{
It will issue a warning if the state number doesn't exist or if there is no data available for the requested
state and year
}
\examples{
fars_map_state(1, "2013")

}
