% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file and return a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to read}
}
\value{
A tibble of the data in the file
}
\description{
This function takes a cvs file, checks that it exists, and reads it, returning a tibble.
}
\details{
It uses reader::read_csv and any warning messages from readr::read_csv will be suppressed
but the function will issue an error if the file doesn't exist
}
\examples{
fars_read(make_filename("2013"))

}
