% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads data from the bz compressed accidents files that correspond to the list of years passed as a parameter.
Returns a list of tibbles with data for each month for each year requested.
Prints a warning if a requested year isn't found in the available data files from the Fatality Analysis Reporting System}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector or list of years}
}
\value{
list of tibbles with data for each year from the file accidents_<year>.csv.bz2,
plus attributes for the data, such as the state, county and others'
}
\description{
Reads data from the bz compressed accidents files that correspond to the list of years passed as a parameter.
Returns a list of tibbles with data for each month for each year requested.
Prints a warning if a requested year isn't found in the available data files from the Fatality Analysis Reporting System
}
\examples{
fars_read_years(c("2013","2015"))

}
