% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function returns the total number of accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector or list of years}
}
\value{
Tibble with the summarized data by month with result for each year in its own column
}
\description{
This function takes in a list of years for which data should be summarized.
}
\examples{
fars_summarize_years(c("2013","2015"))

}
