% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Create a string to use as a filename for bz2 compressed csv file for accident data in the Fatality Analysis Reporting System.}
\usage{
make_filename(year)
}
\arguments{
\item{Year}{to use a suffix in the filename}
}
\value{
String in format accident_<year>/csv/bz2
}
\description{
This function takes the year and returns a string in the format accident_<year>.csv.bz2
}
\examples{
make_filename("2013")

}
