% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_plot_data.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the state number to be used}

\item{year}{the year to be plotted}
}
\value{
a map of the selected state with the accident data plotted
}
\description{
Generates a map of the state selected with the accident locations plotted using the latitude & longitude data for #'the selected year
}
\details{
filename: fars_plot_data.R
}
\examples{
fars_map_state(30, 2013)


Error conditions:
fars_map_state(3,2013)
> Error in fars_map_state(3, 2013) : invalid STATE number: 3

}
