% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data_load.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A filename (and path to file from working directory if required) to be read in using the function.}
}
\value{
This function returns a dataframe tibble to the console.
}
\description{
Reads in file to data variable using the read_csv function and creates a data frame tibble summarising the #'contents.
}
\details{
filename: fars_data_load.R

These functions read in data taken from the US National Highway Traffic Safety Administration's
[Fatality Analysis Reporting System](https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)).
}
\examples{
fars_read("directory/filename")
fars_read("data/accident_2013.csv.bz2")

Error conditions:
\dontrun {fars_read(missing_file)}

Incorrect filepath or missing file will result in "file 'filename' does not exist" error.

}
