% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data_load.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{year(s) to be summarized}
}
\value{
a dataframe with the data sorted by year and month.
}
\description{
Reads the selected file into a dataframe then groups the data by year and month. Uses make_filename function to #'allow the user to select the year for the data rather than inputting the whole filename.
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013:2015))

Error conditions: Invalid year will result in error.
fars_read_years(2001)
>[[1]]
NULL

Warning message:
 In value[[3L]](cond) : invalid year: 2001

}
