% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_data.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{selected year(s) to summarize.}
}
\value{
A list of integers representing the months and accident counts for the selected year.
}
\description{
Summarizes accident data by month for selcted year(s).
}
\details{
filename: fars_summarize_data.R

This function summarises data taken from the US National Highway Traffic Safety Administration's
[Fatality Analysis Reporting System](https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)).
The fars_data_load functions (make_filename, fars_read & fars_read_years) import the data.
}
\examples{
fars_summarize_years(2013)

Error conditions: Years where there are no data will result in error.
\dontrun {fars_summarize_years(3100)}

Error in grouped_df_impl(data, unname(vars), drop) :
Column `year` is unknown

}
