% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data_load.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{for the filename.}
}
\value{
character vector
}
\description{
Helper function. Reads input, converts it to an interger and then creates a character vector incorporating
the input value to use as a filename using the sprintf function.
}
\examples{
make_filename(2001)
> "accident_2001.csv.bz2"

Error conditions:
make_filename(tulip)

non numeric inputs will cause errors.

}
